+++ src/PortalBase/CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `\Heptacom\HeptaConnect\Portal\Base\Mapping\MappedDatasetEntityCollection::groupByPortalNode` as usage of `\spl_object_hash` is discouraged and grouping can be solved in storage implementations more efficiently
 - Move deprecated `\Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PathMethodsTrait` contents into `\Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PackageContract`
 - Remove composer dependency `opis/closure` as it is no longer needed with the removal of closure deferrals as it has not been a practical solution to defer closure execution in a different process
+- Remove `\Heptacom\HeptaConnect\Portal\Base\Portal\Contract\PortalNodeContextInterface::getPortal` without replacement. Use dependency-injection if you need the portal instance.
 
 ### Fixed